Francionando quantidades em OS e Vendas

+++ application/libraries/Gateways/Asaas.php
@@ -187,14 +187,14 @@ class Asaas extends BasePaymentGateway
         $totalProdutos = array_reduce(
             $produtos,
             function ($total, $item) {
-                return $total + (floatval($item->preco) * intval($item->quantidade));
+                return $total + (floatval($item->preco) * floatval($item->quantidade));
             },
             0
         );
         $totalServicos = array_reduce(
             $servicos,
             function ($total, $item) {
-                return $total + (floatval($item->preco) * intval($item->quantidade));
+                return $total + (floatval($item->preco) * floatval($item->quantidade));
             },
             0
         );
@@ -302,14 +302,14 @@ class Asaas extends BasePaymentGateway
         $totalProdutos = array_reduce(
             $produtos,
             function ($total, $item) {
-                return $total + (floatval($item->preco) * intval($item->quantidade));
+                return $total + (floatval($item->preco) * floatval($item->quantidade));
             },
             0
         );
         $totalServicos = array_reduce(
             $servicos,
             function ($total, $item) {
-                return $total + (floatval($item->preco) * intval($item->quantidade));
+                return $total + (floatval($item->preco) * floatval($item->quantidade));
             },
             0
         );
