+++ application/libraries/Gateways/GerencianetSdk.php
@@ -177,14 +177,14 @@ class GerencianetSdk extends BasePaymentGateway
         $totalProdutos = array_reduce(
             $produtos,
             function ($total, $item) {
-                return $total + (floatval($item->preco) * intval($item->quantidade));
+                return $total + (floatval($item->preco) * floatval($item->quantidade));
             },
             0
         );
         $totalServicos = array_reduce(
             $servicos,
             function ($total, $item) {
-                return $total + (floatval($item->preco) * intval($item->quantidade));
+                return $total + (floatval($item->preco) * floatval($item->quantidade));
             },
             0
         );
@@ -324,14 +324,14 @@ class GerencianetSdk extends BasePaymentGateway
         $totalProdutos = array_reduce(
             $produtos,
             function ($total, $item) {
-                return $total + (floatval($item->preco) * intval($item->quantidade));
+                return $total + (floatval($item->preco) * floatval($item->quantidade));
             },
             0
         );
         $totalServicos = array_reduce(
             $servicos,
             function ($total, $item) {
-                return $total + (floatval($item->preco) * intval($item->quantidade));
+                return $total + (floatval($item->preco) * floatval($item->quantidade));
             },
             0
         );
